+++ src/components/MessageList/utils.ts
@@ -8,6 +8,7 @@ import { isDate } from '../../context/TranslationContext';
 import type { MessageLabel, UserResponse } from 'stream-chat';
 import type { DefaultStreamChatGenerics } from '../../types/types';
 import type { StreamMessage } from '../../context/ChannelStateContext';
+import { isMessageEdited } from '../Message/utils';
 
 type ProcessMessagesParams<
   StreamChatGenerics extends DefaultStreamChatGenerics = DefaultStreamChatGenerics
@@ -280,7 +281,8 @@ export const getGroupStyles = <
     message.user?.id !== previousMessage.user?.id ||
     previousMessage.type === 'error' ||
     previousMessage.deleted_at ||
-    (message.reaction_counts && Object.keys(message.reaction_counts).length > 0);
+    (message.reaction_counts && Object.keys(message.reaction_counts).length > 0) ||
+    isMessageEdited(previousMessage);
 
   const isBottomMessage =
     !nextMessage ||
@@ -291,7 +293,8 @@ export const getGroupStyles = <
     message.user?.id !== nextMessage.user?.id ||
     nextMessage.type === 'error' ||
     nextMessage.deleted_at ||
-    (nextMessage.reaction_counts && Object.keys(nextMessage.reaction_counts).length > 0);
+    (nextMessage.reaction_counts && Object.keys(nextMessage.reaction_counts).length > 0) ||
+    isMessageEdited(message);
 
   if (!isTopMessage && !isBottomMessage) {
     if (message.deleted_at || message.type === 'error') return 'single';